--- a/signac/job.py
+++ b/signac/job.py
@@ -12,6 +12,7 @@ import shutil
 from copy import deepcopy
 from threading import RLock
 from typing import FrozenSet
+from types import MappingProxyType
 
 from synced_collections.backends.collection_json import (
     BufferedJSONAttrDict,
@@ -287,12 +288,12 @@ class Job:
             except TypeError:
                 raise KeyTypeError
 
-            self._statepoint_dict = deepcopy(statepoint)
+            self._statepoint_mapping = statepoint
             self._statepoint_requires_init = True
         else:
             # Only an id was provided. State point will be loaded lazily.
             self._id = id_
-            self._statepoint_dict = None
+            self._statepoint_mapping = None
             self._statepoint_requires_init = True
 
     def _initialize_lazy_properties(self):
@@ -331,7 +332,7 @@ class Job:
 
     def __repr__(self):
         return "{}(project={}, statepoint={})".format(
-            self.__class__.__name__, repr(self._project), self.statepoint_dict
+            self.__class__.__name__, repr(self._project), self.statepoint_mapping
         )
 
     @property
@@ -404,10 +405,10 @@ class Job:
         self.statepoint = statepoint
 
     @property
-    def statepoint_dict(self):
-        """Get a copy of the job's statepoint as a dict.
+    def statepoint_mapping(self):
+        """Get a copy of the job's statepoint as a read-only mapping.
 
-        `statepoint_dict` uses the statepoint cache to provide fast access to the
+        `statepoint_mapping` uses the statepoint cache to provide fast access to the
         job's statepoint for reading.
 
         .. note::
@@ -421,13 +422,13 @@ class Job:
 
         Returns
         -------
-        dict
+        Mapping
             Returns the job's state point.
         """
-        if self._statepoint_dict is None:
-            self._statepoint_dict = deepcopy(self._project._get_statepoint(self._id))
+        if self._statepoint_mapping is None:
+            self._statepoint_mapping = self._project._get_statepoint(self._id)
 
-        return self._statepoint_dict
+        return MappingProxyType(self._statepoint_mapping)
 
     @property
     def statepoint(self):
@@ -441,7 +442,7 @@ class Job:
 
         .. tip::
 
-            Use `statepoint_dict` for faster access to read the statepoint.
+            Use `statepoint_mapping` for faster access to read the statepoint.
 
         .. warning::
 
@@ -470,21 +471,19 @@ class Job:
         """
         with self._lock:
             if self._statepoint_requires_init:
-                if self._statepoint_dict is None:
+                if self._statepoint_mapping is None:
                     # Load state point data lazily (on access).
                     self._statepoint = _StatePointDict(
                         jobs=[self], filename=self._statepoint_filename,
                     )
                     statepoint = self._statepoint.load(self.id)
 
-                    self._statepoint_dict_requires_init = False
-
                     # Update the project's state point cache when loaded lazily
                     self._project._register(self.id, statepoint)
                 else:
                     # Create _StatePointDict lazily with a known statepoint dict.
                     self._statepoint = _StatePointDict(
-                        jobs=[self], filename=self._statepoint_filename, data=self._statepoint_dict
+                        jobs=[self], filename=self._statepoint_filename, data=self._statepoint_mapping
                     )
 
                 self._statepoint_requires_init = False