<!-- Provide a general summary of your changes in the title above. -->

## Description
<!-- Describe your changes in detail. -->
<!-- Please indicate if the changes may break existing functionality. -->
Improve the performance of signac for many use-cases, especially when workspaces have with large job counts. Read the complete commit messages for full details on each change.

To summarize:
* Introduce `Job.statepoint_mapping` (edit: `Job.cached_statepoint`)which allows cached, read-only access to the statepoint. `statepoint_mapping` is loaded on demand when the job is not in the cache.
* Use the statepoint cache in additional code paths.
* All `open_job` code paths now lazily load `Job._statepoint`.
* Add `validate_statepoint` argument to `Job.init`. When `False`, `init` checks only that the job directory exists.
* Cache job ids in `JobsCursor` so that `__len__` and `__contains__` are O(1).
* Re-use the results from `listdir` when iterating over jobs in `Project` and bypass the `exists` check on every job as it is opened.

## Motivation and Context
<!-- Why is this change required? What problem does it solve? -->
<!-- If it fixes an open issue, please link to the issue here. -->
Users would like their scripts to complete quickly. I will post benchmark results in the comments.

## Checklist:
<!-- This checklist must be complete before merging the pull request. -->
<!-- If you are unsure about any of these items, do not hesitate to ask! -->
- [x] I am familiar with the [Contributing Guidelines](https://github.com/glotzerlab/signac/blob/main/CONTRIBUTING.md).
- [x] I agree with the terms of the [Contributor Agreement](https://github.com/glotzerlab/signac/blob/main/ContributorAgreement.md).
- [x] My name is on the [list of contributors](https://github.com/glotzerlab/signac/blob/main/contributors.yaml).
- [x] The changes introduced by this pull request are covered by existing or newly introduced tests.
- [ ] The [package documentation](https://github.com/glotzerlab/signac/tree/main/doc) and [framework documentation](https://docs.signac.io/) in [signac-docs](https://github.com/glotzerlab/signac-docs) are up to date with these changes.
- [x] I have updated the [changelog](https://github.com/glotzerlab/signac/blob/main/changelog.txt) and added any related issue and pull request numbers for future reference.