@joaander 2024-02-07
@tcmoore3 @janbridley here are the signac modifications I've been talking about. I will post benchmarks soon.