@joaander 2024-02-08
Here are the results for a suite of benchmarks run on cheme-hodges (NVME) with 100,000 jobs in the workspace. For example, the `access_statepoint_mapping` benchmark is:
```python
for job in project:
    job.statepoint_mapping['a']
```

The "no cache" column is run with no statepoint cache file on disk. `cached` is measured after executing `signac update-cache`.

`main`:
| Benchmark                 | no cache (s)   | cached (s)   |
|---------------------------|----------------|--------------|
| iterate_jobs              | 0.365          | 0.379        |
| with_job                  | 5.6            | 5.65         |
| open_job_id               | 0.299          | 0.303        |
| open_job_statepoint       | 1.34           | 1.31         |
| open_job_statepoint_id    | 0.942          | 0.938        |
| access_statepoint_mapping | 3.73           | 3.72         |
| access_statepoint         | 3.55           | 3.62         |
| access_job_document       | 8.08           | 8.03         |
| access_job_stores         | 34.2           | 33.9         |
| find_all                  | 2.07           | 1.36         |
| groupby                   | 2.47           | 1.66         |
| project_len               | 6.6            | 6.62         |
| project_contains          | 0.47           | 0.474        |
| find_jobs_len             | 24.9           | 30.1         |
| find_jobs_contains        | >240           | >240         |

This pull request (7998f9a0b2084f99493312d17ee28391cffce8e4):
| Benchmark                 |   no cache (s) |   cached (s) |
|---------------------------|----------------|--------------|
| iterate_jobs              |         0.158  |       0.16   |
| with_job                  |         0.533  |       0.541  |
| open_job_id               |         0.323  |       0.169  |
| open_job_statepoint       |         0.507  |       0.579  |
| open_job_statepoint_id    |         0.0733 |       0.0762 |
| access_statepoint_mapping |         1.08   |       0.292  |
| access_statepoint         |         3.27   |       3.3    |
| access_job_document       |         3.01   |       2.79   |
| access_job_stores         |        25.5    |      25.6    |
| find_all                  |         1.15   |       0.478  |
| groupby                   |         2.54   |       1.65   |
| project_len               |         6.64   |       6.67   |
| project_contains          |         0.497  |       0.398  |
| find_jobs_len             |         1.06   |       0.391  |
| find_jobs_contains        |         1.17   |       0.527  |