@joaander 2024-02-08
Here is the same on Great Lakes scratch (GPFS).

`main`:
| Benchmark                 | no cache (s)   | cached (s)   |
|---------------------------|----------------|--------------|
| iterate_jobs              | 45.5           | 46.6         |
| with_job                  | 136.0          | 128.0        |
| open_job_id               | 49.0           | 50.6         |
| open_job_statepoint       | 1.69           | 1.66         |
| open_job_statepoint_id    | 1.19           | 1.17         |
| access_statepoint_mapping | 117.0          | 114.0        |
| access_statepoint         | 117.0          | 123.0        |
| access_job_document       | 199.0          | 201.0        |
| access_job_stores         | >240           | >240         |
| find_all                  | 120.0          | 2.17         |
| groupby                   | 113.0          | 2.24         |
| project_len               | 9.37           | 9.36         |
| project_contains          | 50.1           | 52.3         |
| find_jobs_len             | 151.0          | 45.1         |
| find_jobs_contains        | >240           | >240         |

This pull request (7998f9a0b2084f99493312d17ee28391cffce8e4):
| Benchmark                 | no cache (s)   | cached (s)   |
|---------------------------|----------------|--------------|
| iterate_jobs              | 0.214          | 0.215        |
| with_job                  | 52.5           | 53.2         |
| open_job_id               | 48.8           | 0.256        |
| open_job_statepoint       | 0.616          | 0.732        |
| open_job_statepoint_id    | 0.0893         | 0.0877       |
| access_statepoint_mapping | 116.0          | 0.403        |
| access_statepoint         | 110.0          | 115.0        |
| access_job_document       | 116.0          | 111.0        |
| access_job_stores         | >240           | >240         |
| find_all                  | 119.0          | 0.764        |
| groupby                   | 115.0          | 2.23         |
| project_len               | 9.42           | 9.41         |
| project_contains          | 50.8           | 50.7         |
| find_jobs_len             | 112.0          | 0.66         |
| find_jobs_contains        | 106.0          | 0.763        |