@joaander 2024-02-08
This is ready for review. I suggest waiting to merge until I complete additional testing of this branch with signac-flow.