@joaander 2024-02-08
All changes around `statepoint_mapping` and the more aggressive use of the statepoint cache fully retain the existing data model. As noted in one of the commit comments, the only significant change occurs when a user opens a job by statepoint with numpy arrays in the dictionary. The `NumpyConversion` warning is deferred until/if the user access `Job.statepoint`. (They would have received this warning when previously calling Job.`init()`).

Removing the exists checks while iterating and caching ids in `find_jobs` is crucial to allowing faster iteration through the project. Depending on timing, this *may* (or may not) cause different behavior in cases where users remove job directories after starting iteration but before the iteration reaches the removed job. In signac 2.1, the user would receive a `KeyError`: https://github.com/glotzerlab/signac/blob/2e9a23b3c781de5a1a3ca915f9177cdf75e78286/signac/project.py#L545-L547 In this pull request, the user would receive a `Job` instance that knows its id at a minumim. If the statepoint is in the cache, the user will be able to call `job.init()` to restore the missing directory. If the statepoint is not in the cache, then the user will get an error when calling `job.init()`. There are similar race conditions if the user attempts to open the job document or stores after the job is created if the job directory is removed.

Iterating over jobs while at the same time removing them is asking for undefined behavior. I don't see a functional difference between a `KeyError` from *signac* or e.g. a `FileNotExists` error when a user attempts to open `job.fn('some_file')`.

Please be aware that these race conditions existed prior to this pull request. A user could remove the directory after the `self._contains_job_id(id)` check but before the job is accessed (similarly for documents and stores) and get the same `FileNotExists` error as with this pull request.

I don't see how this PR changes durability. signac 2.1.0 already lazily loads the statepoint in many cases. Any invalid modifications to `signac_statepoint.json` would previously have been detected only when it is loaded.