@joaander 2024-02-10
This pull request breaks flow `aggregate.groupsof` (and possibly other aggregates). The jobs appear in random orders in the aggregates. I am investigating.